Добавление JWT-токена для сессии Реализация своих методов БД

+++ app/users/heandlers.py
@@ -3,9 +3,9 @@ from starlette import status
 from fastapi import APIRouter, Body, Depends, HTTPException
 from app.users.schemas import UserLoginForm, CreateUserForm, UpdateUserForm
 from app.models import connection_db, User, AuthToken
-from app.utils import get_password_hash
+from app.utils import get_password_hash, get_session_token
 from app.auth import check_auth_token
-
+from app.db_methods.methods import Database
 
 router = APIRouter()
 
@@ -16,56 +16,54 @@ async def index():
 
 
 @router.post('/user/login', name='login to service')
-def login(user_form: UserLoginForm = Body(..., embed=True), database=Depends(connection_db)):
+def login(user_form: UserLoginForm = Body(..., embed=True), database=Depends(Database)):
     """Login to user account"""
-    user = database.query(User).filter(User.email == user_form.email).one_or_none()
+    user = database.session.query(User, AuthToken).join(AuthToken, AuthToken.user_id == User.user_id).filter(User.email == user_form.email).one_or_none()
+    # TODO доделать запрос к бд
     if not user or get_password_hash(user_form.password) != user.password:
         return {'error': 'Email or password invalid'}
 
-    auth_token = AuthToken(token=str(uuid.uuid4()), user_id=user.user_id)
-    token = auth_token.token
-    try:
-        database.add(auth_token)
-        database.commit()
-    except Exception as _ex:
-        return {'error': _ex}
-    finally:
-        database.close()
-    return {'auth_token': f'{token}'}
+    # user_token = database.session.query(AuthToken).filter(AuthToken.user_id == user.user_id).one_or_none()
+    if user.token is not None:
+        # return {'status':'Account already authorized'})
+        return {'session_token': f'{user.token}'}
+
+    session_token = AuthToken(token=get_session_token(user.user_id), user_id=user.user_id)
+    db_resp = database.add(session_token)
+    if db_resp:
+        return {'session_token': f'{session_token.token}'}
+    HTTPException(status_code=status.HTTP_501_NOT_IMPLEMENTED, detail=db_resp)
 
 
 @router.post('/user/logout', name='logout from service')
-def logout(token: AuthToken = Depends(check_auth_token), database=Depends(connection_db)):
+def logout(token: AuthToken = Depends(check_auth_token), database=Depends(Database)):
     """Removing a user from the database by authorization token (to protect against removal from outside)"""
-    user = database.query(User).filter(User.user_id == token.user_id).one()
+    user = database.session.query(User).filter(User.user_id == token.user_id).one()
     email = user.email
-    try:
-        database.delete(token)
-        database.commit()
-    except Exception as _ex:
-        return {'error': _ex}
-    finally:
-        database.close()
-
-    return {f'user {email}: Logout'}
+    db_resp = database.delete(token)
+    if db_resp:
+        return {f'user {email}: Logout'}
+    HTTPException(status_code=status.HTTP_501_NOT_IMPLEMENTED, detail=db_resp)
 
 
-@router.get('/user/info')
-def get_user(token: AuthToken = Depends(check_auth_token), database=Depends(connection_db)):
+@router.post('/user/info')
+def get_user(token: AuthToken = Depends(check_auth_token), database=Depends(Database)):
     """Getting user information by his token"""
-    user = database.query(User).filter(User.user_id == token.user_id).one_or_none()
-    return {'id': user.user_id, 'email': user.email, 'nickname': user.nickname}
+    user = database.session.query(User).filter(User.user_id == token.user_id).one_or_none()
+    if user:
+        return {'id': user.user_id, 'email': user.email, 'nickname': user.nickname}
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='User not found')
 
 
 @router.post('/user/create', name='user:create', response_model_exclude_unset=True)
-def create_user(user: CreateUserForm = Body(..., embed=True), database=Depends(connection_db)):
+def create_user(user: CreateUserForm = Body(..., embed=True), database=Depends(Database)):
     """To generate the correct password, the following must be used:
     1 digit,
     1 capital letter,
     1 small letter,
     Password length: from 8 to 20 characters,
     Use only latin letters"""
-    exists_user = database.query(User.user_id).filter(User.email == user.email).one_or_none()
+    exists_user = database.session.query(User.user_id).filter(User.email == user.email).one_or_none()
     if exists_user:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail='Email already exists')
 
@@ -74,52 +72,46 @@ def create_user(user: CreateUserForm = Body(..., embed=True), database=Depends(c
         password=get_password_hash(user.password),
         nickname=user.nickname
     )
-    user_id = new_user.user_id
-    try:
-        database.add(new_user)
-        database.commit()
-    except Exception as _ex:
-        return {'error': _ex}
-    finally:
-        database.close()
-
-    return {'User': f'Created user {user_id}'}
+    db_resp = database.add(new_user)
+    if db_resp:
+        return {'User': f'Created user {user.email}'}
+    raise HTTPException(status_code=status.HTTP_501_NOT_IMPLEMENTED, detail=db_resp)
 
 
-@router.put('/user', response_model_exclude_unset=True)
-def update_user(new_user: UpdateUserForm, token: AuthToken = Depends(check_auth_token), database=Depends(connection_db)):
+@router.post('/user/update', response_model_exclude_unset=True)
+def update_user(new_user: UpdateUserForm, old_pwd: str = Body(..., embed=True),
+                token: AuthToken = Depends(check_auth_token), database=Depends(Database)):
     """Update account information. The correctness of the password and email address is also checked.
     Parameters are optional, can be changed by choice"""
     user = database.query(User).filter(User.user_id == token.user_id).one()
-    if new_user.email:
-        user.email = new_user.email
-    if new_user.password:
-        user.password = get_password_hash(new_user.password)
-    if new_user.nickname:
-        user.nickname = new_user.nickname
-    try:
-        database.add(user)
-        database.commit()
-    except Exception as _ex:
-        return {'error': _ex}
-    finally:
-        database.close()
-    return {'status': "Successfully"}
+    if user.password == get_password_hash(old_pwd):
+        if new_user.email:
+            user.email = new_user.email
+        if new_user.password:
+            user.password = get_password_hash(new_user.password)
+        if new_user.nickname:
+            user.nickname = new_user.nickname
+        db_resp = database.add(user)
+        if db_resp:
+            return {'status': "Successfully"}
+        raise HTTPException(status_code=status.HTTP_501_NOT_IMPLEMENTED, detail=db_resp)
+    return {'error': 'Wrong old password!'}
 
 
-@router.delete('/user')
-def delete_user(token: AuthToken = Depends(check_auth_token), database=Depends(connection_db)):
+@router.post('/user/delete')
+def delete_user(old_pwd: str = Body(..., embed=True), token: AuthToken = Depends(check_auth_token),
+                database=Depends(Database)):
     """Removing a user from the database by authorization token (to protect against removal from outside)"""
-    user = database.query(User).filter(User.user_id == token.user_id).one()
+    user = database.session.query(User).filter(User.user_id == token.user_id).one()
     if user:
         email = user.email
-        try:
-            database.delete(user)
-            database.commit()
-        except Exception as _ex:
-            return {'error': _ex}
-        finally:
-            database.close()
-        return {f'user {email}: Deleted'}
+        if get_password_hash(old_pwd) == user.password:
+            db_resp_token = database.delete(token)
+            db_resp_user = database.delete(user)
+            if db_resp_token and db_resp_user:
+                return {f'user {email}: Deleted'}
+            raise HTTPException(status_code=status.HTTP_501_NOT_IMPLEMENTED,
+                                detail=(not db_resp_token or not db_resp_user))
+        return {'error': 'Wrong old password!'}
     return {'answer': 'This user does not exist'}
